config(nixos/machines): Update `/boot` partition UUID
Previously `/boot` on this machine was only `100 MiB`. Next to this
partition was the Windows Recovery Environment partition taking extra
`500 MiB`. Since I don't care about recovering Windows, I deleted WinRE
in order to enlarge `/boot` (which was very often suffering from
insufficient space during `nixos-rebuild switch`). I tried to do that
with GParted, but it seems that I hit this bug:

    https://bugzilla.gnome.org/show_bug.cgi?id=649324

So I followed the workaround to backup `/boot` and recreate it again,
but that caused the UUID of the partition to change.

This commit updates the `fileSystems."/boot"` option with the new UUID.